fix refreshing, other refactor
