+++ composeApp/src/commonMain/kotlin/org/yaabelozerov/gotovomp/data/network/RecipePagingSource.kt
@@ -1,7 +1,9 @@
 package org.yaabelozerov.gotovomp.data.network
 
+import androidx.paging.Pager
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
+import app.cash.paging.PagingConfig
 import org.yaabelozerov.gotovomp.data.network.model.RecipeDTO
 import org.yaabelozerov.gotovomp.domain.otherwise
 import org.yaabelozerov.gotovomp.domain.unwrap
@@ -14,17 +16,23 @@ class RecipePagingSource(
     }
 
     override suspend fun load(params: LoadParams<Int>): LoadResult<Int, RecipeDTO> {
-        val nextPage = params.key ?: 0
+        val nextPage = params.key ?: 1
         var res: LoadResult<Int, RecipeDTO>? = null
-        client.getRecipeFeed(limit = 5, offset = nextPage).unwrap { resp ->
+        client.getRecipeFeed(limit = PAGE_SIZE, offset = nextPage - 1).unwrap { resp ->
             res = LoadResult.Page(
                 data = resp,
                 prevKey = null,
-                nextKey = (nextPage + 5).takeIf { resp.isNotEmpty() }
+                nextKey = (nextPage - 1 + PAGE_SIZE).takeIf { resp.isNotEmpty() }
             )
         } otherwise {
             res = LoadResult.Error(Throwable(it.toString()))
         }
         return res!!
     }
+
+    companion object {
+        const val PAGE_SIZE = 5
+    }
 }
+
+fun createPager(apiClient: ApiClient) = Pager(PagingConfig(RecipePagingSource.PAGE_SIZE)) { RecipePagingSource(apiClient) }
