+++ composeApp/src/commonMain/kotlin/org/yaabelozerov/gotovomp/domain/usecase/FeedUseCase.kt
@@ -1,22 +1,15 @@
-package org.yaabelozerov.gotovomp.domain.use_cases
+package org.yaabelozerov.gotovomp.domain.usecase
 
-import androidx.paging.Pager
-import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 import org.yaabelozerov.gotovomp.data.network.ApiClient
-import org.yaabelozerov.gotovomp.data.network.RecipePagingSource
+import org.yaabelozerov.gotovomp.data.network.createPager
 import org.yaabelozerov.gotovomp.data.network.model.RecipeDTO
 
 class FeedUseCase(
     private val api: ApiClient
 ) {
-    fun getFeedFlow(): Flow<PagingData<RecipeDTO>> {
-        val pager = Pager(PagingConfig(pageSize = 1)) {
-            RecipePagingSource(api)
-        }.flow
-        return pager
-    }
+    fun getFeedFlow(): Flow<PagingData<RecipeDTO>> = createPager(api).flow
 
     fun likeRecipe() {
         TODO("Not yet implemented")
@@ -25,5 +18,4 @@ class FeedUseCase(
     fun commentRecipe(text: String) {
         TODO("Not yet implemented")
     }
-
 }
