+++ composeApp/src/commonMain/kotlin/org/yaabelozerov/gotovomp/presentation/components/AlternativeRecipeCard.kt
@@ -58,121 +58,7 @@ import org.yaabelozerov.gotovomp.presentation.util.tertiaryDark
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
-fun AlternativeRecipeCard(
-    recipe: RecipeDTO,
-    imageUrl: String,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var imageLoaded by rememberSaveable { mutableStateOf(false) }
-    val fadeColor =
-        MaterialTheme.colorScheme.run { if (!isSystemInDarkTheme()) onBackground else background }
-    val fadeInverse =
-        MaterialTheme.colorScheme.run { if (isSystemInDarkTheme()) onBackground else background }
-
-    Card(
-        modifier = modifier.fillMaxWidth().clickable { onClick() },
-        colors = CardDefaults.outlinedCardColors(
-            containerColor = MaterialTheme.colorScheme.primaryContainer,
-            contentColor = MaterialTheme.colorScheme.primary
-        )
-    ) {
-        Column(Modifier.fillMaxWidth()) {
-            Row(
-                modifier = Modifier.fillMaxWidth().padding(8.dp),
-                horizontalArrangement = Arrangement.spacedBy(6.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Box(Modifier.size(36.dp).clip(CircleShape).background(color = tertiaryDark))
-                Text(
-                    text = recipe.author.run { firstName?.plus(" ")?.plus(lastName) ?: username },
-                    style = MaterialTheme.typography.titleMedium
-                )
-            }
-            Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.BottomStart) {
-                var lineCount by remember { mutableStateOf(1) }
-                AsyncImage(
-                    modifier = Modifier.fillMaxWidth().aspectRatio(0.6f).drawWithContent {
-                        drawContent()
-                        drawRect(
-                            brush = Brush.verticalGradient(
-                                0f to Color.Transparent,
-                                0.65f - .15f*lineCount to Color.Transparent,
-                                1f - .15f*lineCount  to fadeColor.copy(0.85f)
-                            )
-                        )
-                    },
-                    model = imageUrl,
-                    contentDescription = recipe.name,
-                    contentScale = ContentScale.Crop,
-                    onSuccess = { imageLoaded = true })
-                Text(
-                    modifier = Modifier.padding(16.dp),
-                    text = recipe.name,
-                    overflow = TextOverflow.Ellipsis,
-                    maxLines = 3,
-                    style = MaterialTheme.typography.displayMedium.copy(
-                        lineBreak = LineBreak.Paragraph,
-                        color = fadeInverse,
-                        fontWeight = FontWeight.Black
-                    ),
-                    onTextLayout = {
-                        lineCount = it.lineCount
-                    },
-                )
-            }
-            Column(modifier = Modifier.padding(vertical = 12.dp), verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text(
-                    "Ингредиенты:",
-                    modifier = Modifier.padding(horizontal = 12.dp),
-                )
-                var expandIngredients by remember { mutableStateOf(false) }
-                AnimatedContent(expandIngredients) { expanded ->
-                    if (expanded) {
-                        recipe.ingredients.run {
-                            FlowRow(horizontalArrangement = Arrangement.spacedBy(8.dp), verticalArrangement = Arrangement.spacedBy(8.dp),
-                                modifier = Modifier.padding(horizontal = 12.dp).clickable { expandIngredients = !expandIngredients }) {
-                                forEach {
-                                    OutlinedCard(shape = MaterialTheme.shapes.large, colors = CardDefaults.outlinedCardColors(containerColor = fadeInverse, contentColor = fadeColor)) {
-                                        Text(it.name, modifier = Modifier.padding(horizontal = 12.dp, vertical = 4.dp), fontStyle = FontStyle.Italic)
-                                    }
-                                }
-                            }
-                        }
-
-                    } else {
-                        LazyRow(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.clickable { expandIngredients = !expandIngredients }) {
-                            item {
-                                Spacer(Modifier.width(4.dp))
-                            }
-                            items(recipe.ingredients.take(7)) {
-                                OutlinedCard(shape = MaterialTheme.shapes.large, colors = CardDefaults.outlinedCardColors(containerColor = fadeInverse, contentColor = fadeColor)) {
-                                    Text(it.name, modifier = Modifier.padding(horizontal = 12.dp, vertical = 4.dp), fontStyle = FontStyle.Italic)
-                                }
-                            }
-                            item {
-                                Spacer(Modifier.width(4.dp))
-                            }
-                        }
-                    }
-                }
-            }
-            Row(modifier = Modifier.fillMaxWidth().background(color = MaterialTheme.colorScheme.surfaceContainerHigh), horizontalArrangement = Arrangement.spacedBy(12.dp), verticalAlignment = Alignment.CenterVertically) {
-                IconButton({}, modifier = Modifier.padding(start = 4.dp)) { Icon(Icons.Outlined.ThumbUp, contentDescription = "Like") }
-                IconButton({}) { Icon(Icons.Outlined.Add, contentDescription = "Comment") }
-                IconButton({}) { Icon(Icons.Outlined.Share, contentDescription = "Share") }
-                Spacer(Modifier.weight(1f))
-                Text("час назад", style = MaterialTheme.typography.labelMedium)
-                Spacer(Modifier.size(8.dp))
-
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-fun AlternativeRecipeCard2(
+fun RecipeCard(
     recipe: RecipeDTO,
     imageUrl: String,
     onClick: () -> Unit,
