+++ composeApp/src/commonMain/kotlin/org/yaabelozerov/gotovomp/viewModels/FeedScreenVM.kt
@@ -2,16 +2,16 @@ package org.yaabelozerov.gotovomp.viewModels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.paging.cachedIn
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.yaabelozerov.gotovomp.domain.DomainError
-import org.yaabelozerov.gotovomp.data.network.ApiClient
 import org.yaabelozerov.gotovomp.data.network.model.RecipeDTO
 import org.yaabelozerov.gotovomp.domain.GlobalEvent
 import org.yaabelozerov.gotovomp.domain.UI
-import org.yaabelozerov.gotovomp.domain.use_cases.FeedUseCase
+import org.yaabelozerov.gotovomp.domain.usecase.FeedUseCase
 
 data class MainScreenVMState(
     val recipes: List<RecipeDTO> = emptyList(),
@@ -20,13 +20,12 @@ data class MainScreenVMState(
 )
 
 class FeedScreenVM(
-    private val client: ApiClient,
-    private val useCase: FeedUseCase = FeedUseCase(client)
+    useCase: FeedUseCase
 ) : ViewModel() {
     private val _state = MutableStateFlow(MainScreenVMState())
     val state = _state.asStateFlow()
 
-    val recipes = useCase.getFeedFlow()
+    val recipes = useCase.getFeedFlow().cachedIn(viewModelScope)
 
     init {
         subscribeToGlobalEvents()
